@@ -1,6 +1,5 @@
-// @ts-ignore: No type information available yet https://github.com/waitandseeagency/graphql-type-datetime/issues/5
-import GraphQLDateTime from 'graphql-type-datetime'
 import { mergeResolvers } from '@graphql-tools/merge'
+import { DateTimeResolver, EmailAddressResolver } from 'graphql-scalars'
 import { Resolvers } from './graphql'
 import { resolvers as userResolvers } from './user/resolvers'
 import { resolvers as documentResolvers } from './documents/resolvers'
@@ -13,7 +12,8 @@ export function loadResolvers(): Resolvers {
     groupResolvers(),
     {
       // Custom scalar types
-      DateTime: GraphQLDateTime,
+      DateTime: DateTimeResolver,
+      EmailAddress: EmailAddressResolver,
     },
   ])
 }
