@@ -4,6 +4,7 @@ import session from 'express-session'
 import passport from 'passport'
 import { RedisClient } from 'redis'
 import { injectable } from 'tsyringe'
+import { config } from '../../config'
 import { AuthService } from './auth.service'
 import EmailStrategy from './auth.email.strategy'
 
@@ -20,7 +21,7 @@ export default class PassportInitializer {
       this.serializeUser(user, done)
     )
     passport.deserializeUser<string>((id, done) =>
-      this.deserializeUser(id as string, done)
+      this.deserializeUser(id, done)
     )
   }
 
@@ -35,7 +36,7 @@ export default class PassportInitializer {
           disableTouch: true,
         }),
         // The secret used to sign the session cookie
-        secret: 'TODO: CHANGE THIS TO ENVIRONMENT VARIABLE',
+        secret: [config.session.primarySecret, config.session.secondarySecret],
         // Don't force session to be saved back to the session store unless it was modified
         resave: false,
         saveUninitialized: false,
@@ -60,19 +61,24 @@ export default class PassportInitializer {
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   private serializeUser(user: any, done: (err: unknown, id?: string) => void) {
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
     done(null, user.id)
   }
 
-  private async deserializeUser(
+  private deserializeUser(
     id: string,
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     done: (err: unknown, user?: any) => void
   ) {
-    const user = await this.accountService.getUserById(id)
-    if (user === undefined) {
-      done("account doesn't exist", undefined)
-    } else {
-      done(null, user)
-    }
+    this.accountService
+      .getUserById(id)
+      .then((user) => {
+        if (user === undefined) {
+          done("account doesn't exist", undefined)
+        } else {
+          done(null, user)
+        }
+      })
+      .catch((error) => done(error, null))
   }
 }
