@@ -50,15 +50,14 @@
           <FontAwesomeIcon icon="times" />
         </t-button>
         <slot>
-          <div class="flex flex-col">
+          <div class="flex flex-col overflow-y-auto pr-3">
             <div>
               <t-select
                 :options="{
                   Article: 'Journal Article',
                   PhDThesis: 'PhD Thesis',
                 }"
-                class="text-gray-500 focus-within:text-black"
-                variant="plain"
+                variant="plaincaps"
               >
                 <template slot="arrow" slot-scope="{ className }">
                   <FontAwesomeIcon
@@ -69,7 +68,7 @@
                 </template>
               </t-select>
             </div>
-            <div class="z-10 grid">
+            <div class="z-10 grid -mt-2">
               <!-- Auto-grow textarea, taken from https://css-tricks.com/the-cleanest-trick-for-autogrowing-textareas/ -->
               <!-- prettier-ignore -->
               <div
@@ -97,14 +96,76 @@
                 "
               ></t-textarea>
             </div>
-            <div>
+            <div class="-mt-3">
               <Tags
-                ref="myRef"
                 v-model="authors"
                 placeholder="Add author"
                 :delimiters="null"
                 :whitelist="authorSuggestions"
               />
+            </div>
+            <div>
+              <editor-input v-model="date"></editor-input>
+              <span class="text-gray-400">|</span>
+              <editor-input v-model="journal"></editor-input>
+            </div>
+            <div class="-mt-1">
+              <span class="pl-3">
+                Volume:
+                <editor-input v-model="volume"></editor-input>
+              </span>
+              <span class="text-gray-400 pr-3">|</span>
+              <span>
+                Issue:
+                <editor-input v-model="issue"></editor-input>
+              </span>
+              <span class="text-gray-400 pr-3">|</span>
+              <span>
+                pp.
+                <editor-input v-model="pages"></editor-input>
+              </span>
+            </div>
+            <div>
+              <editor-header
+                heading="Abstract"
+                class="mt-4 -mb-1"
+              ></editor-header>
+              <t-textarea
+                v-model="abstract"
+                variant="plain"
+                rows="5"
+              ></t-textarea>
+            </div>
+            <div>
+              <editor-header heading="Keywords" class="mt-4"></editor-header>
+              <Tags
+                v-model="keywords"
+                placeholder="Add keyword"
+                :delimiters="null"
+                :whitelist="keywordSuggestions"
+                tag-class="border rounded-md"
+              />
+            </div>
+            <div>
+              <editor-header heading="Groups" class="mt-4"></editor-header>
+              <Tags
+                v-model="groups"
+                placeholder="Add group"
+                :delimiters="null"
+                :whitelist="groupSuggestions"
+                tag-class="bg-highlight-50 rounded-md"
+              />
+            </div>
+            <div>
+              <editor-header
+                heading="External"
+                class="mt-4 mb-1"
+              ></editor-header>
+              <t-table
+                :data="external"
+                variant="plain"
+                class="text-sm"
+              ></t-table>
             </div>
           </div>
         </slot>
@@ -116,17 +177,38 @@
 import { defineComponent, computed } from '@nuxtjs/composition-api'
 import Tags from './tagify.vue'
 import { useUiStore } from './../store'
+import EditorInput from './EditorInput.vue'
+import EditorHeader from './EditorHeader.vue'
 
 export default defineComponent({
   components: {
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
     Tags,
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+    EditorInput,
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+    EditorHeader,
   },
   setup() {
     const ui = useUiStore()
 
     const authors = [{ value: 'Tobias Diez' }, { value: 'Gerd Rudolph' }]
-
     const authorSuggestions = [{ value: 'Linus' }]
+
+    const keywords = [
+      { value: 'Symplectic Geometry' },
+      { value: 'Yang-Mills Theory' },
+    ]
+    const keywordSuggestions = [{ value: 'Differential Geometry' }]
+
+    const groups = [{ value: 'Chocolate Making' }, { value: 'Consumption' }]
+    const groupSuggestions = [{ value: 'Grinding' }]
+
+    const external = [
+      { name: 'DOI', value: '10.1063/1.5085764' },
+      { name: 'ArXiv', value: '1812.04695' },
+      { name: 'MathSciNet', value: 'MR3997558' },
+    ]
 
     return {
       isDetailsOpen: computed(() => ui.isDetailsOpen),
@@ -135,6 +217,18 @@ export default defineComponent({
         'Clebsch-Lagrange variational principle and geometric constraint analysis of relativistic field theories',
       authors,
       authorSuggestions,
+      date: '2019 / 12',
+      journal: 'Journal of Mathematical Physics',
+      volume: '300',
+      issue: '3',
+      pages: '567-589',
+      abstract:
+        'Inspired by the Clebsch optimal control problem, we introduce a new variational principle that is suitable for capturing the geometry of relativistic field theories with constraints related to a gauge symmetry. Its special feature is that the Lagrange multipliers couple to the configuration variables via the symmetry group action. The resulting constraints are formulated as a condition on the momentum map of the gauge group action on the phase space of the system. We discuss the Hamiltonian picture and the reduction of the gauge symmetry by stages in this geometric setting. We show that the Yang-Mills-Higgs action and the Einstein--Hilbert action fit into this new framework after a (1+3)-splitting. Moreover, we recover the Gauß constraint of Yang-Mills-Higgs theory and the diffeomorphism constraint of general relativity as momentum map constraints.',
+      keywords,
+      keywordSuggestions,
+      groups,
+      groupSuggestions,
+      external,
     }
   },
 })
