@@ -13,46 +13,49 @@
         ></FontAwesomeIcon>
         <span>{{ document.title }}</span>
       </button>
-
+      <!-- TOOD: Add citation display
       <a
         class="text-sm whitespace-nowrap mt-1"
-        href="/paper/3104609148/citedby"
+        href="/paper/id/citedby"
       >
         2 citations
       </a>
+      -->
     </div>
-    <div class="space-x-1">
-      <a href="author/2549788765/publication?paperId=3104609148">Tobias Diez</a>
-      <a href="author/2488101096/publication?paperId=3104609148"
-        >Gerd Rudolph</a
-      >
+    <div v-if="'authors' in document && document.authors" class="space-x-3">
+      <span v-for="author in document.authors" :key="author.id">{{
+        author.name
+      }}</span>
     </div>
     <div class="space-x-1 text-sm">
       <span class="font-semibold">2019</span>
       <a
         v-if="'journal' in document && document.journal"
-        href="journal/158241587"
+        href="journal/TODOid"
         >{{ document.journal.name }}</a
       >
     </div>
-    <div class="flex flex-row space-x-2 text-sm">
-      <t-tag variant="badge" class="border border-gray-400">
-        Diffeomorphism constraint
+    <div
+      v-if="document.keywords.length > 0"
+      class="flex flex-row space-x-2 text-sm"
+    >
+      <t-tag
+        v-for="keyword in document.keywords"
+        :key="keyword"
+        variant="badge"
+        class="border border-gray-400"
+      >
+        <!-- TODO: Add icon of group <FontAwesomeIcon icon="dragon" size="xs" class="mr-2" /> -->
+        {{ keyword }}
       </t-tag>
-      <t-tag variant="badge" class="border border-gray-400">
-        Gauge symmetry
-      </t-tag>
-      <t-tag variant="badge" class="border border-gray-400">
-        <FontAwesomeIcon icon="dragon" size="xs" class="mr-2" />
-        Dragons
-      </t-tag>
-
+      <!-- TODO: Add overflow
       <t-button variant="linkplain" class="text-sm my-auto">
         <span>View More (8+)</span>
         <FontAwesomeIcon icon="chevron-down" size="xs" />
       </t-button>
+      -->
     </div>
-    <div v-if="document.abstract">
+    <div v-if="'abstract' in document && document.abstract">
       <span class="flex-grow" :class="{ 'line-clamp-2': !viewFullAbstract }">
         {{ document.abstract }}
       </span>
@@ -82,12 +85,38 @@ import {
   computed,
   toRefs,
 } from '@nuxtjs/composition-api'
-import { Article, Document, DocumentType } from '../apollo/graphql'
+import { gql } from '@apollo/client/core'
+import { DocumentForViewFragment } from '../apollo/graphql'
 import { useUiStore } from '~/store'
+
+export const DocumentForView = gql`
+  fragment DocumentForView on Document {
+    id
+    title
+    keywords
+    ... on Article {
+      abstract
+      authors {
+        ... on Person {
+          id
+          name
+        }
+        ... on Organization {
+          id
+          name
+        }
+      }
+      journal {
+        name
+      }
+    }
+  }
+`
+
 export default defineComponent({
   props: {
     document: {
-      type: Object as PropType<Document | Article>,
+      type: Object as PropType<DocumentForViewFragment>,
       required: true,
     },
   },
@@ -95,16 +124,17 @@ export default defineComponent({
     const { document } = toRefs(props)
     const viewFullAbstract = ref(false)
 
+    // TODO: Convert the following switches to use `document.value.type`
     const typeIcon = computed(() => {
-      switch (document.value.type) {
-        case DocumentType.Article:
+      switch (document.value.__typename) {
+        case 'Article':
           return 'newspaper'
       }
     })
 
     const typeDescription = computed(() => {
-      switch (document.value.type) {
-        case DocumentType.Article:
+      switch (document.value.__typename) {
+        case 'Article':
           return 'Journal Paper'
       }
     })
