@@ -1,24 +1,30 @@
 <template>
   <div>
-    <Portal to="header">
-      <Logo class="mx-auto h-20 w-auto" />
-      <h2 class="mt-8 text-center text-5xl font-extrabold text-gray-900">
-        Create account
-      </h2>
-      <p class="mt-8 text-center text-sm text-gray-600">
-        Already have an account?
-        <t-nuxtlink to="/user/login">Sign in</t-nuxtlink>
-      </p>
-      <t-alert
-        v-if="error"
-        variant="error"
-        class="mt-8"
-        :dismissible="false"
-        show
-      >
-        {{ error }}
-      </t-alert>
+    <Portal to="side">
+      <div class="flex flex-col">
+        <img
+          class="w-11/12 mx-auto transform flip-horizontal"
+          src="~/assets/undraw_road_to_knowledge_m8s0.svg"
+        />
+        <div class="mt-7 mx-auto text-2xl">One last step!</div>
+      </div>
     </Portal>
+    <h2 class="text-center text-5xl font-extrabold text-gray-900">
+      Create account
+    </h2>
+    <p class="mt-6 mb-8 text-center text-sm text-gray-600">
+      Already have an account?
+      <t-nuxtlink to="/user/login">Sign in</t-nuxtlink>
+    </p>
+    <t-alert
+      v-if="error"
+      variant="error"
+      class="mt-8"
+      :dismissible="false"
+      show
+    >
+      {{ error }}
+    </t-alert>
     <form @submit.prevent="signup">
       <div class="space-y-5">
         <t-input-group label="Email address" variant="important">
@@ -50,9 +56,10 @@
   </div>
 </template>
 <script lang="ts">
-import gql from 'graphql-tag'
+import { gql } from '@apollo/client/core'
 import { defineComponent, ref, useRouter } from '@nuxtjs/composition-api'
-import { useSignupMutation } from '~/apollo/graphql'
+import { useMutation } from '@vue/apollo-composable'
+import { SignupDocument } from '~/apollo/graphql'
 import { currentUserVar } from '~/apollo/cache'
 
 export default defineComponent({
@@ -64,7 +71,7 @@ export default defineComponent({
     const password = ref('')
 
     gql`
-      mutation signup($email: String!, $password: String!) {
+      mutation signup($email: EmailAddress!, $password: String!) {
         signup(email: $email, password: $password) {
           id
         }
@@ -74,7 +81,7 @@ export default defineComponent({
       mutate: signup,
       onDone,
       error,
-    } = useSignupMutation(() => ({
+    } = useMutation(SignupDocument, () => ({
       variables: {
         email: email.value,
         password: password.value,
@@ -85,7 +92,7 @@ export default defineComponent({
     }))
     const router = useRouter()
     onDone(() => {
-      router.push('/dashboard')
+      void router.push('/dashboard')
     })
 
     return {
