@@ -2,16 +2,14 @@ import { PrismaClient, GroupType, GroupHierarchyType } from '@prisma/client'
 
 async function seedInternal(prisma: PrismaClient): Promise<void> {
   await prisma.user.deleteMany({})
-  // Manually delete other fields first (at the moment, its not possible to do this automatically https://github.com/prisma/prisma/issues/2810)
-  await prisma.userDocumentOtherField.deleteMany({})
   await prisma.userDocument.deleteMany({})
   await prisma.group.deleteMany({})
 
   await prisma.user.create({
     data: {
       id: 'ckn4oul7100004cv7y3t94n8j',
-      email: `test@testum.de2`,
-      password: `$2a$10$d.P8dAo.0pmsvW2xwsH7sudRMhSBiCvOfEkN8JBeuvHBs1ZtLUdoe`, // EBNPXY35TYkYXHs
+      email: 'test@testum.de2',
+      password: '$2a$10$d.P8dAo.0pmsvW2xwsH7sudRMhSBiCvOfEkN8JBeuvHBs1ZtLUdoe', // EBNPXY35TYkYXHs
       name: 'Alice',
     },
   })
